methods dont work

+++ src/main/java/com/example/cinemaApp/models/User.java
@@ -9,7 +9,7 @@ public class User implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column
     private String username;
@@ -54,8 +54,8 @@ public class User implements Serializable {
         this.active = active;
     }
 
-    public User(Integer id, String username, String password, String name, String surname, String role,
-            String phoneNumber, String email, String dateOfBirth, Boolean active) {
+    public User(Long id, String username, String password, String name, String surname, String role, String phoneNumber,
+            String email, String dateOfBirth, Boolean active) {
         this.id = id;
         this.username = username;
         this.password = password;
@@ -68,11 +68,11 @@ public class User implements Serializable {
         this.active = active;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
